+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgraded dependencies
 - Add back button on advanced settings page
 - Sliver app bar in entity definitions
+- Sliver app bar in speech settings
 
 ## [0.8.355] - 2023-05-19
 ### Added:
